extract sharding.rs
